+++ components/src/main/java/com/mercadolibre/android/andesui/radiobutton/type/AndesRadioButtonTypeInterface.kt
@@ -18,7 +18,7 @@ internal object AndesRadioButtonTypeIdle : AndesRadioButtonTypeInterface {
         return if (status == AndesRadioButtonStatus.UNSELECTED) {
             R.color.andes_gray_250_solid.toAndesColor()
         } else {
-            R.color.andes_blue_ml_500.toAndesColor()
+            R.color.andes_accent_color_500.toAndesColor()
         }
     }
     override fun iconColor(context: Context, status: AndesRadioButtonStatus) = R.color.andes_white.toAndesColor()
@@ -26,7 +26,7 @@ internal object AndesRadioButtonTypeIdle : AndesRadioButtonTypeInterface {
         return if (status == AndesRadioButtonStatus.UNSELECTED) {
             R.color.andes_white.toAndesColor()
         } else {
-            R.color.andes_blue_ml_500.toAndesColor()
+            R.color.andes_accent_color_500.toAndesColor()
         }
     }
     override fun textColor(context: Context) = R.color.andes_gray_800_solid.toAndesColor()