+++ components/src/test/java/com/mercadolibre/android/andesui/textfieldcode/AndesTextfieldCodeA11yTest.kt
@@ -1,0 +1,47 @@
+package com.mercadolibre.android.andesui.textfieldcode
+
+import android.content.Context
+import android.os.Build
+import android.widget.EditText
+import androidx.test.core.app.ApplicationProvider
+import com.facebook.common.logging.FLog
+import com.facebook.drawee.backends.pipeline.Fresco
+import com.facebook.imagepipeline.core.ImagePipelineConfig
+import com.facebook.imagepipeline.listener.RequestListener
+import com.facebook.imagepipeline.listener.RequestLoggingListener
+import com.facebook.soloader.SoLoader
+import com.mercadolibre.android.andesui.textfield.AndesTextfieldCode
+import junit.framework.Assert.assertEquals
+import org.junit.Before
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.annotation.Config
+
+@RunWith(RobolectricTestRunner::class)
+@Config(sdk = [Build.VERSION_CODES.LOLLIPOP])
+class AndesTextfieldCodeA11yTest {
+
+    private lateinit var context: Context
+    private lateinit var textfieldCode: AndesTextfieldCode
+
+    @Before
+    fun setup() {
+        context = ApplicationProvider.getApplicationContext()
+        SoLoader.setInTestMode()
+        val requestListeners = setOf<RequestListener>(RequestLoggingListener())
+        val config = ImagePipelineConfig.newBuilder(context)
+                // other setters
+                .setRequestListeners(requestListeners)
+                .build()
+        Fresco.initialize(context, config)
+        FLog.setMinimumLoggingLevel(FLog.VERBOSE)
+    }
+
+    @Test
+    fun `call accessibilityClassName returns correct value`() {
+        textfieldCode = AndesTextfieldCode(context)
+        val role = EditText::class.java.name
+        assertEquals(role, textfieldCode.accessibilityClassName)
+    }
+}