+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # v3.18.0
 ## 🚀 Feature
+- Added full accessibility support for `AndesTextfield` and `AndesTextfieldCode`
 - Added Indeterminate support to `AndesDropdownForm` | Author: [@MKiperszmid](https://github.com/MKiperszmid)
+## 🛠 Fixes
+- Fix AndesRadioButton color to andes_accent_color_500 | Author: [@gimena-martinez](https://github.com/gimena-martinez)
 
 # v3.17.1
 ## 🛠 Fixes